+++ container/hashmap.h
@@ -28,11 +28,48 @@
 #include <bcplib/core/bcplib_base.h>
 #include <bcplib/container/map.h>
 #include <bcplib/container/arraylist.h>
+/*----- Begin of class hashmap entry -----*/
 BEGIN_DEFINE_CLASS(hashmap_entry)
 #define hashmap_entry_prototype     map_entry_prototype;\
                                     hashmap_entry next;\
                                     size_t hash
 END_DEFINE_CLASS(hashmap_entry)
+/*----- Begin of class hashmap iterator -----*/
+BEGIN_DEFINE_CLASS(hashmap_iterator)
+#define hashmap_iterator_prototype  iterator_prototype;\
+                                    hashmap_entry next_ent;\
+                                    size_t next_slot
+        
+bool hashmap_iterator_has_next(id self);
+void* hashmap_iterator_next(id self);
+void hashmap_iterator_remove(id self);
+END_DEFINE_CLASS(hashmap_iterator)
+/*----- Begin of class hashmap key iterator -----*/
+BEGIN_DEFINE_CLASS(hashmap_key_iterator)
+#define hashmap_key_iterator_prototype    hashmap_iterator_prototype
+void* hashmap_key_iterator_next(id self);
+END_DEFINE_CLASS(hashmap_key_iterator)
+/*----- Begin of class hashmap value iterator -----*/
+BEGIN_DEFINE_CLASS(hashmap_value_iterator)
+#define hashmap_value_iterator_prototype  hashmap_iterator_prototype
+void* hashmap_value_iterator_next(id self);
+END_DEFINE_CLASS(hashmap_value_iterator)        
+/*----- Begin of class hashmap entry set -----*/
+BEGIN_DEFINE_CLASS(hashmap_entry_set)
+#define hashmap_entry_set_prototype     set_prototype;\
+                                        hashmap _host
+iterator hashmap_entry_set_create_iterator(id self);
+
+END_DEFINE_CLASS(hashmap_entry_set)
+/*----- Begin of class hashmap key set -----*/
+BEGIN_DEFINE_CLASS(hashmap_key_set)
+#define hashmap_key_set_prototype       set_prototype
+END_DEFINE_CLASS(hashmap_key_set)
+/*----- Begin of class hashmap values -----*/
+BEGIN_DEFINE_CLASS(hashmap_values)
+#define hashmap_values_prototype        collection_prototype
+END_DEFINE_CLASS(hashmap_values)
+/*----- Begin of class hashmap -----*/
 BEGIN_DEFINE_CLASS(hashmap)
 #define HASHMAP_INIT_COMPARATOR 1
 #define HASHMAP_INIT_HASHER (HASHMAP_INIT_COMPARATOR << 1)